fix(uart-cobs): fix CRC-16 table corruption, O_RDWR descriptor, timesync threshold, and safe spec parsing

+++ bpv7/uart-cobs/uartcobsclo.c
@@ -64,7 +64,7 @@ int main(int argc, char *argv[])
 
 		int bytesSent = sendBundleByUartCobs(&hostNbr, &ductSocket, bundleLength, bundleZco, buffer);
 		
-		if (bytesSent < bundleLength) {
+		if (bytesSent < (int)bundleLength) {
 			sm_SemEnd(uartcloSemaphore(NULL));
 			continue;
 		}
